Changed input to generative functions
Previously, the generative functions used in fitting, n_binomials and binom_isotope, took in an integer (the # of exchangable deuterons) as their first input. But now, these take in an array from 0 to the total # of exchangable deuterons instead. I've chnaged these functions, as well as all instances when they are called, accordingly.
The rationale for this is that now, it will be easier to convert spurious peaks to nans
In my filter spurious peaks function I will change both the y values to Nans rather than 0s at the spuriosu peaks, then use the ignore nan policy during curve fitting. The ignore nan policy didn't work back when I was inputting an integer as my x range